split into multiple files + explicit access control
